gateway/pyproject.toml
[project]
name = "gateway"
version = "0.0.1"
requires-python = "==3.13.2"
dependencies = [
    "loguru==0.7.3",
    "FastAPI==0.115.13",
    "colorama==0.4.6",
    "tabulate==0.9.0",
    "chardet==5.2.0",
    "python-dotenv==1.1.0",
    "grpcio-tools==1.72.1",
    "uvicorn==0.34.3",
    "strawberry-graphql==0.274.0",
]
show_params_on_run = true
# debug_mode parameter is declaring in .env file!


# [tool.ruff]
# line-length = 88


[tool.ruff.lint]
extend-select = ["W", "N", "R", "I"]
# Letters means included rules classes from PEP8
ignore = ["W293"]


[fastapi_server]
fastapi_host = "0.0.0.0"
fastapi_port = 8080


[grpc_server]
max_workers = 15
host = "0.0.0.0"
port = 8081
reload = true
log_requests = true
log_responses = false
log_huge_reqs_and_resps = false
replace_huge_logs_by_small_msgs = true
count_of_chars_in_huge_logs = 200